Refactor location of ChildList; add types to test.
