+++ client/components/edit/ObjectEditor.test.tsx
@@ -10,7 +10,7 @@ jest.mock("../../context/EditorContext", () => ({
         return mockUseEditorContext();
     },
 }));
-jest.mock("./ChildList", () => () => "ChildList");
+jest.mock("./children/ChildList", () => () => "ChildList");
 jest.mock("./Breadcrumbs", () => () => "Breadcrumbs");
 jest.mock("./datastream/DatastreamList", () => () => "DatastreamList");
 jest.mock("./datastream/DatastreamModal", () => () => "DatastreamModal");
